refactor(client,server): creating new scoped  http controller MyProfile UserProfile
