+++ packages/client/services/UsersSearchService.ts
@@ -44,7 +44,7 @@ export async function fetchUsers(
 export async function getMyProfileInformation(
     body: GetMyProfileInformationRequestBody,
 ): Promise<GetMyProfileInformationResponseBody> {
-    const url = urlcat(SERVER_ENDPOINT, '/user/my-profile-information');
+    const url = urlcat(SERVER_ENDPOINT, '/me/profile-information');
 
     const rawResponse = await redaxios.post(url, body);
     const parsedBody = GetMyProfileInformationResponseBody.parse(
